separate appwindow imp module
